content/base/test/test_bug868999.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=868999
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 868999</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=868999">Mozilla Bug 869006</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 868999 **/

var r = new Range();
ok(r, "Range has been created");

var doc = document.implementation.createDocument("", "", null);
var h1 = doc.createElement('h1');
doc.appendChild(h1);

var t = doc.createTextNode('Hello world');
h1.appendChild(t);

r.selectNodeContents(doc);
is(r.toString(), "Hello world", "new Range() works!");

</script>
</pre>
</body>
</html>
